feat: csuc i gas: refactored
