+++ example/curvy-os/index.html
@@ -180,7 +180,7 @@
                                 <label for="fee">Fee:</label>
                                 <input type="text" disabled id="csuc-onboard-fee" value="-" />
                             </div>
-                            <button onClick="onboardIntoCSUC()" id="csuc-onboard-button">Onboard</button>
+                            <button onClick="onboardToCSUC()" id="csuc-onboard-button">Onboard</button>
                         </div>
 
                         <div id ="csuc-transfer-content" role="sub-tab-content">
