+++ example/curvy-os/package.json
@@ -1,12 +1,13 @@
 {
     "name": "curvy-os",
-    "version": "1.0.0",
+    "version": "1.0.1",
     "description": "",
     "main": "index.js",
     "type": "module",
     "scripts": {
+        "prebuild": "cd ../../ && pnpm build",
         "test": "echo \"Error: no test specified\" && exit 1",
-        "start": "pnpm build && vite",
+        "start": "vite",
         "build": "vite build"
     },
     "keywords": [],
