+++ src/rpc/abstract.ts
@@ -43,11 +43,4 @@ export default abstract class RPC {
     abstract CreateReferenceToERC20(token: Address): Promise<any>;
 
     abstract CreateReferenceToCSUC(): Promise<any>;
-
-    abstract PrepareTransferIntoCSUC(
-        from: CurvyStealthAddress,
-        to: `0x${string}`,
-        token: `0x${string}`,
-        amount: bigint
-    ): Promise<any>;
 }
