+++ src/types.ts
@@ -96,42 +96,3 @@ export type AuthConfig = {
     apiKey?: string;
     bearerToken?: string;
 };
-
-// -------------------------------- Gas Sponsorship
-
-export type CreateGasSponsorshipRequest = {
-    actions: GasSponsorship.Action[];
-};
-
-export type CreateGasSponsorshipResponse = {
-    actionIds: string[];
-};
-
-export namespace GasSponsorship {
-    export type Action = {
-        id?: string;
-        networkId: number;
-        payloads: ActionPayload[];
-        signedPayloads: string[];
-    };
-
-    export type ActionPayload = {
-        data: string;
-    };
-
-    export enum ActionStatus {
-        INVALID = "INVALID",
-        ACCEPTED = "ACCEPTED",
-        BATCHED = "BATCHED",
-        FINALIZED = "FINALIZED",
-    }
-
-    export type OnchainData = {
-        fundingTxHash: string;
-        fundingTxBlockHash?: string | null;
-        approveTxHash: string;
-        approveTxBlockHash?: string | null;
-        csucWrappedTxHash: string;
-        csucWrappedTxBlockHash?: string | null;
-    };
-}
